Improved the meaning of the "last played" value in the scoring. The value now ranges from -x to 0 where x is the value of the scoring_lastPlayed setting-variable. If the song has just been played it's value is -x. This value decreases linearly during one week (604800 seconds) at which point it reaches 0. This means that a song that has not been played for a week is not affected by this value.
One could add another condition that favors songs that have not been played for
a very long time. The reason: It is likely that in a database with many songs,
will only play through the subset of songs that fall into the rating mechanism
(those that have been played at least 10 times).

Example case: Let's take a song that has not been played for, say 4 weeks, but
has been played already 5 times. This song is not bossted for being new, and
the "last played" score has reached 0 as well already. So the rating-score is 0
and the lastplayed score is 0) Which gives it a total score of 0 (assuming the
"added-to-database" score is also 0). This put's it in front of songs that have
been played within a week and have not been rated (as those scores are
negative). But songs that have been rated might have a positive score. So a
song that has a rating of 0.1 for example and has not been played for 2 weeks,
is still preferred to the unrated songs.

A solution for this problem has to be found. One possibility is to give
non-rated songs the score of 0.5 instead of 0. That would make more sense
anyhow, as the ratings rate from 0 (skipped reaaaaaly often), and 1 (never
skipped). So even a song that everybody keeps skipping has a higher score than
those never even rated (given it has been played to the end at least once).

Actually, let's do the "0.5" thing right now.

brb....



git-svn-id: http://svn.foobar.lu/wickedjukebox/trunk@129 f24607ba-0e10-0410-b95a-ef2f3848dab9